api/requirements.txt
absl-py==0.2.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
Django==2.0.6
django-admin==1.2.6
django-excel-response2==2.0.8
django-filter==2.0.0
django-six==1.0.4
djangorestframework==3.8.2
dlib==19.13.99
enum34==1.1.6
face-recognition==1.2.2
face-recognition-models==0.3.0
future==0.16.0
idna==2.7
Markdown==2.6.11
mock==2.0.0
mpyq==0.2.5
numpy==1.14.5
opencv-python==3.4.1.15
pbr==4.0.4
Pillow==5.1.0
portpicker==1.2.0
protobuf==3.6.0
pygame==1.9.3
PySC2==2.0.1
pytz==2018.4
requests==2.19.1
s2clientprotocol==4.3.2.65384.0
scipy==1.1.0
screen==1.0.1
six==1.11.0
sk-video==1.1.10
urllib3==1.23
websocket-client==0.48.0
whichcraft==0.4.1
xlwt==1.3.0
